Implement place query methods and tests
Currently the tests which check for a null result are failing because
the actual API is returning an empty list rather than null - unlike for
the postcodes query endpoint. Have submitted an issue for this with
PostcodesIO (https://memb.li/2V48r1X) and will update this if needed
in due course.